Fixed integration tests.
