+++ src/RabbitMQ.Client.Core.DependencyInjection/InternalExtensions/RabbitMqServiceOptionsDependencyInjectionExtensions.cs
@@ -16,11 +16,8 @@ namespace RabbitMQ.Client.Core.DependencyInjection.InternalExtensions
             return options;
         }
 
-        internal static IServiceCollection ConfigureRabbitMqProducingServiceOptions(this IServiceCollection services, RabbitMqServiceOptions options) =>
+        internal static IServiceCollection ConfigureRabbitMqProducingOnlyServiceOptions(this IServiceCollection services, RabbitMqServiceOptions options) =>
             services.Configure<RabbitMqConnectionOptions>(x => x.ProducerOptions = options);
-
-        internal static IServiceCollection ConfigureRabbitMqConsumingServiceOptions(this IServiceCollection services, RabbitMqServiceOptions options) =>
-            services.Configure<RabbitMqConnectionOptions>(x => x.ConsumerOptions = options);
 
         internal static IServiceCollection ConfigureRabbitMqConnectionOptions(this IServiceCollection services, RabbitMqServiceOptions options) =>
             services.Configure<RabbitMqConnectionOptions>(x => { x.ProducerOptions = options; x.ConsumerOptions = options; });
