+++ src/RabbitMQ.Client.Core.DependencyInjection/Services/Interfaces/IChannelDeclarationService.cs
@@ -1,0 +1,8 @@
+namespace RabbitMQ.Client.Core.DependencyInjection.Services.Interfaces
+{
+    // TODO: summary.
+    public interface IChannelDeclarationService
+    {
+        void SetConnectionInfrastructureForRabbitMqServices();
+    }
+}
